requirements.txt
compas_model
compas_viewer
compas_notebook
git+https://github.com/BRG-research/compas_grid.git